A project for the exam of complex systems
This is a project for the exam of complex systems. It simulates the
spreading pattern of mobile
phone  virus  by  means  of  SMS  spam  and  Bluetooth.  By  varying
all  the
parameters  of  correspondent  model  and  analyzing  the  different
results,  the
study will find out how each parameter will influence the propagation of
mobile
phone virus.